Add a new UDB Logo command: EXEC()
This is like DO() in that it calls a subroutine, but instead of pushing the subroutine onto the logo stack, this subroutine becomes the top of the logo stack.  When/if it returns, the UDB Logo program starts over from the beginning.  This is useful when calling multiple non-terminating subroutines remotely, and you want to avoid overrunning the logo stack.