Serial IO Improvements and Cleanup
Changes all Serial communication to be asynchronous.  This includes all config messages sent to the GPS, HILSIM data sent to the GPS port, and nmea-passthrough for Remzibi OSD support.

Use -1 instead of 0 as the EOF value for serial output for telemetry and GPS.

This fixes the new glitch on output channel 3 when running the RollPitchYaw Demo, that had been visible on some setups.

Also, only tangentially relatedly, changes the format used to upload a waypoint and a logo command to be distinct from eachother.